Add Command refactor
