+++ test/pact/command/exec_command_test.exs
@@ -42,9 +42,7 @@ defmodule Kadena.Pact.ExecCommandTest do
 
       %{
         meta_data: MetaData.new(raw_meta_data),
-        raw_meta_data: raw_meta_data,
         keypair: KeyPair.new(keypair_data),
-        keypair_data: keypair_data,
         signer:
           Signer.new(
             pub_key: pub_key,
@@ -159,72 +157,6 @@ defmodule Kadena.Pact.ExecCommandTest do
         |> ExecCommand.build()
     end
 
-    test "with a valid keyword list in meta_data", %{
-      raw_meta_data: raw_meta_data,
-      keypair: keypair,
-      signer: signer,
-      nonce: nonce,
-      code: code
-    } do
-      %Command{
-        cmd:
-          "{\"meta\":{\"chainId\":\"0\",\"creationTime\":1667249173,\"gasLimit\":1000,\"gasPrice\":1.0e-6,\"sender\":\"k:554754f48b16df24b552f6832dda090642ed9658559fef9f3ee1bb4637ea7c94\",\"ttl\":28800},\"networkId\":\"testnet04\",\"nonce\":\"2023-06-13 17:45:18.211131 UTC\",\"payload\":{\"exec\":{\"code\":\"(+ 5 6)\",\"data\":{}}},\"signers\":[{\"addr\":\"6ffea3fabe4e7fe6a89f88fc6d662c764ed1359fbc03a28afdac3935415347d7\",\"clist\":[{\"args\":[\"6ffea3fabe4e7fe6a89f88fc6d662c764ed1359fbc03a28afdac3935415347d7\"],\"name\":\"coin.GAS\"}],\"pubKey\":\"6ffea3fabe4e7fe6a89f88fc6d662c764ed1359fbc03a28afdac3935415347d7\",\"scheme\":\"ED25519\"}]}",
-        hash: %PactTransactionHash{
-          hash: "3BeOVboVQ9TwWAiQS65YHkgxOOiBQT5uMcQDm3vmNVA"
-        },
-        sigs: %SignaturesList{
-          signatures: [
-            %Signature{
-              sig:
-                "e4ea07728a52b564156ea44aa2817931b56829eaa361f1b5c4f2a419acaa5c9c77e4e29ac55f5d250d27d066dc8b4f1076fdc5d402fa4cbffe2f63dc36e1e705"
-            }
-          ]
-        }
-      } =
-        ExecCommand.new()
-        |> ExecCommand.set_network(:testnet04)
-        |> ExecCommand.set_data(%{})
-        |> ExecCommand.set_code(code)
-        |> ExecCommand.set_nonce(nonce)
-        |> ExecCommand.set_metadata(raw_meta_data)
-        |> ExecCommand.add_keypair(keypair)
-        |> ExecCommand.add_signer(signer)
-        |> ExecCommand.build()
-    end
-
-    test "with a valid keyword list in keypair", %{
-      meta_data: meta_data,
-      keypair_data: keypair_data,
-      signer: signer,
-      nonce: nonce,
-      code: code
-    } do
-      %Command{
-        cmd:
-          "{\"meta\":{\"chainId\":\"0\",\"creationTime\":1667249173,\"gasLimit\":1000,\"gasPrice\":1.0e-6,\"sender\":\"k:554754f48b16df24b552f6832dda090642ed9658559fef9f3ee1bb4637ea7c94\",\"ttl\":28800},\"networkId\":\"testnet04\",\"nonce\":\"2023-06-13 17:45:18.211131 UTC\",\"payload\":{\"exec\":{\"code\":\"(+ 5 6)\",\"data\":{}}},\"signers\":[{\"addr\":\"6ffea3fabe4e7fe6a89f88fc6d662c764ed1359fbc03a28afdac3935415347d7\",\"clist\":[{\"args\":[\"6ffea3fabe4e7fe6a89f88fc6d662c764ed1359fbc03a28afdac3935415347d7\"],\"name\":\"coin.GAS\"}],\"pubKey\":\"6ffea3fabe4e7fe6a89f88fc6d662c764ed1359fbc03a28afdac3935415347d7\",\"scheme\":\"ED25519\"}]}",
-        hash: %PactTransactionHash{
-          hash: "3BeOVboVQ9TwWAiQS65YHkgxOOiBQT5uMcQDm3vmNVA"
-        },
-        sigs: %SignaturesList{
-          signatures: [
-            %Signature{
-              sig:
-                "e4ea07728a52b564156ea44aa2817931b56829eaa361f1b5c4f2a419acaa5c9c77e4e29ac55f5d250d27d066dc8b4f1076fdc5d402fa4cbffe2f63dc36e1e705"
-            }
-          ]
-        }
-      } =
-        ExecCommand.new()
-        |> ExecCommand.set_network(:testnet04)
-        |> ExecCommand.set_data(%{})
-        |> ExecCommand.set_code(code)
-        |> ExecCommand.set_nonce(nonce)
-        |> ExecCommand.set_metadata(meta_data)
-        |> ExecCommand.add_keypair(keypair_data)
-        |> ExecCommand.add_signer(signer)
-        |> ExecCommand.build()
-    end
-
     test "with a keypair list", %{
       meta_data: meta_data,
       keypair: keypair,
@@ -297,41 +229,6 @@ defmodule Kadena.Pact.ExecCommandTest do
         |> ExecCommand.build()
     end
 
-    test "with a signers list", %{
-      meta_data: meta_data,
-      keypair: keypair,
-      signer: signer,
-      nonce: nonce,
-      code: code
-    } do
-      signers_list = [signer, signer]
-
-      %Command{
-        cmd:
-          "{\"meta\":{\"chainId\":\"0\",\"creationTime\":1667249173,\"gasLimit\":1000,\"gasPrice\":1.0e-6,\"sender\":\"k:554754f48b16df24b552f6832dda090642ed9658559fef9f3ee1bb4637ea7c94\",\"ttl\":28800},\"networkId\":\"testnet04\",\"nonce\":\"2023-06-13 17:45:18.211131 UTC\",\"payload\":{\"exec\":{\"code\":\"(+ 5 6)\",\"data\":{}}},\"signers\":[{\"addr\":\"6ffea3fabe4e7fe6a89f88fc6d662c764ed1359fbc03a28afdac3935415347d7\",\"clist\":[{\"args\":[\"6ffea3fabe4e7fe6a89f88fc6d662c764ed1359fbc03a28afdac3935415347d7\"],\"name\":\"coin.GAS\"}],\"pubKey\":\"6ffea3fabe4e7fe6a89f88fc6d662c764ed1359fbc03a28afdac3935415347d7\",\"scheme\":\"ED25519\"},{\"addr\":\"6ffea3fabe4e7fe6a89f88fc6d662c764ed1359fbc03a28afdac3935415347d7\",\"clist\":[{\"args\":[\"6ffea3fabe4e7fe6a89f88fc6d662c764ed1359fbc03a28afdac3935415347d7\"],\"name\":\"coin.GAS\"}],\"pubKey\":\"6ffea3fabe4e7fe6a89f88fc6d662c764ed1359fbc03a28afdac3935415347d7\",\"scheme\":\"ED25519\"}]}",
-        hash: %PactTransactionHash{
-          hash: "niwT670sdzjiBWlORejSHsaoVQHPfUiNkxtcVQfDysM"
-        },
-        sigs: %SignaturesList{
-          signatures: [
-            %Signature{
-              sig:
-                "4a35b98eb50179c77784532ea7224d119f06c8a9853c421bd0cd3e667b54213840d01ee926ab19a02d0777f68297b88bbc5799bcd281ed0e9c2c7993dab7740b"
-            }
-          ]
-        }
-      } =
-        ExecCommand.new()
-        |> ExecCommand.set_network(:testnet04)
-        |> ExecCommand.set_data(%{})
-        |> ExecCommand.set_code(code)
-        |> ExecCommand.set_nonce(nonce)
-        |> ExecCommand.set_metadata(meta_data)
-        |> ExecCommand.add_keypair(keypair)
-        |> ExecCommand.add_signers(signers_list)
-        |> ExecCommand.build()
-    end
-
     test "with an invalid data in build function", %{
       meta_data: meta_data,
       keypair: keypair,
@@ -347,7 +244,8 @@ defmodule Kadena.Pact.ExecCommandTest do
         |> ExecCommand.set_nonce(nonce)
         |> ExecCommand.set_metadata(meta_data)
         |> ExecCommand.add_keypair(keypair)
-        |> ExecCommand.add_signers([signer])
+        |> ExecCommand.add_signer(signer)
+        |> ExecCommand.add_signers(signer)
         |> ExecCommand.build()
     end
 
@@ -366,7 +264,7 @@ defmodule Kadena.Pact.ExecCommandTest do
         |> ExecCommand.set_nonce(nonce)
         |> ExecCommand.set_metadata(meta_data)
         |> ExecCommand.add_keypair(keypair)
-        |> ExecCommand.add_signers([signer])
+        |> ExecCommand.add_signer(signer)
     end
 
     test "with an invalid data", %{
@@ -384,7 +282,7 @@ defmodule Kadena.Pact.ExecCommandTest do
         |> ExecCommand.set_nonce(nonce)
         |> ExecCommand.set_metadata(meta_data)
         |> ExecCommand.add_keypair(keypair)
-        |> ExecCommand.add_signers([signer])
+        |> ExecCommand.add_signer(signer)
     end
 
     test "with an invalid code", %{
@@ -402,7 +300,7 @@ defmodule Kadena.Pact.ExecCommandTest do
         |> ExecCommand.set_nonce(nonce)
         |> ExecCommand.set_metadata(meta_data)
         |> ExecCommand.add_keypair(keypair)
-        |> ExecCommand.add_signers([signer])
+        |> ExecCommand.add_signer(signer)
     end
 
     test "with an invalid nonce", %{
@@ -419,7 +317,7 @@ defmodule Kadena.Pact.ExecCommandTest do
         |> ExecCommand.set_nonce(123)
         |> ExecCommand.set_metadata(meta_data)
         |> ExecCommand.add_keypair(keypair)
-        |> ExecCommand.add_signers([signer])
+        |> ExecCommand.add_signer(signer)
     end
 
     test "with an invalid meta_data", %{
@@ -428,7 +326,7 @@ defmodule Kadena.Pact.ExecCommandTest do
       nonce: nonce,
       code: code
     } do
-      {:error, [meta_data: :invalid, args: :not_a_list]} =
+      {:error, [metadata: :invalid]} =
         ExecCommand.new()
         |> ExecCommand.set_network(:testnet04)
         |> ExecCommand.set_data(%{})
@@ -436,7 +334,7 @@ defmodule Kadena.Pact.ExecCommandTest do
         |> ExecCommand.set_nonce(nonce)
         |> ExecCommand.set_metadata("invalid value")
         |> ExecCommand.add_keypairs([keypair])
-        |> ExecCommand.add_signers([signer])
+        |> ExecCommand.add_signer(signer)
     end
 
     test "with an invalid keypair", %{
@@ -445,7 +343,7 @@ defmodule Kadena.Pact.ExecCommandTest do
       nonce: nonce,
       code: code
     } do
-      {:error, [keypair: :invalid, args: :not_a_list]} =
+      {:error, [keypair: :invalid]} =
         ExecCommand.new()
         |> ExecCommand.set_network(:testnet04)
         |> ExecCommand.set_data(%{})
@@ -479,7 +377,7 @@ defmodule Kadena.Pact.ExecCommandTest do
       nonce: nonce,
       code: code
     } do
-      {:error, [keypair: :invalid, args: :not_a_list]} =
+      {:error, [keypair: :invalid]} =
         ExecCommand.new()
         |> ExecCommand.set_network(:testnet04)
         |> ExecCommand.set_data(%{})
@@ -496,7 +394,7 @@ defmodule Kadena.Pact.ExecCommandTest do
       nonce: nonce,
       code: code
     } do
-      {:error, [signers: :invalid, signers: :invalid_type]} =
+      {:error, [signers: :invalid]} =
         ExecCommand.new()
         |> ExecCommand.set_network(:testnet04)
         |> ExecCommand.set_data(%{})
