+++ test/types/meta_data_test.exs
@@ -102,5 +102,9 @@ defmodule Kadena.Types.MetaDataTest do
           chain_id: 0
         )
     end
+
+    test "with an invalid KeywordList" do
+      {:error, [args: :not_a_list]} = MetaData.new("invalid_args")
+    end
   end
 end
